requirements.txt
pandas
Pillow
torch
torchvision
scikit-learn
tqdm
wandb
